Check extra value in received notification.

+++ mobile/android/base/java/org/mozilla/gecko/OuinetBroadcastReceiver.java
@@ -12,19 +12,25 @@ import java.util.Arrays;
 import java.util.List;
 
 public class OuinetBroadcastReceiver extends BroadcastReceiver {
+    // The value constants also force us to use
+    // the right type check for the extras bundle.
     public static final String EXTRA_ACTION_STOP = "org.mozilla.gecko.OuinetBroadcastReceiver.STOP";
+    public static final byte EXTRA_ACTION_STOP_VALUE = 1;
     private static final String TAG = "OuinetBroadcastReceiver";
 
     @Override
     public void onReceive(Context context, Intent intent) {
         Log.d(TAG, "Received intent: "+ intent +", shutting down Ouinet service");
+        if (intent.getExtras().getByte(EXTRA_ACTION_STOP) != EXTRA_ACTION_STOP_VALUE) {
+            return;
+        }
         killPackageProcesses(context);
         Process.killProcess(Process.myPid());
     }
 
     public static Intent createIntent(Context context) {
         Intent intent = new Intent(context, OuinetBroadcastReceiver.class);
-        intent.putExtra(EXTRA_ACTION_STOP, /* unused */ 1);
+        intent.putExtra(EXTRA_ACTION_STOP, EXTRA_ACTION_STOP_VALUE);
         return intent;
     }
